feat(gnoweb): add -style-dir, alternative root and notfound handler, create gnoweb.Options
Main modifications:

- add flag -style-dir <dir>, from which (css|font|img)/ can be read (the
  default styles will still be served as a fallthrough if supplied
  style-dir fails to serve the requested asset).
- add and make use of a gnoweb.Options structure:

+type Options struct {
+	Aliases         map[string]string
+	Redirects       map[string]string
+	RootHandler     func(*slog.Logger, gotuna.App, *Config) http.Handler
+	NotFoundHandler func(*slog.Logger, gotuna.App, *Config) http.Handler
+}

Aliases and Redirects allow to extract the content from the code,
but also to make such mapping entirely customizable and optional.

Root and NotFound handler are custom handlers taking place
of default ones. They allow an entirely different kind of website from
gno.land.

For example, if one does not provide Aliases and Redirects, but
provides a root handler pointing to some markdown file, and
a NotFoundHandler that serves a hierarchy of personal wiki markdown
files, he can mix such a wiki with realms render on top of gno.land.

(see https://github.com/grepsuzette/gnAsteroid for an exploration in
this direction).